librustc_target: Unconfigure tests during normal build
